feat: implement fingerprint injection subsystem
- Added FingerprintApplier class to handle the application of browser fingerprint profiles to tabs and their associated worker targets.
- Refactored Tab class to delegate fingerprint application to FingerprintApplier, simplifying the apply_fingerprint method.
- Enhanced documentation in fingerprint-injection.md to clarify the importance of consistent fingerprinting across layers and added case studies for better understanding.
